requirements.txt
cryptography==42.0.4
fastembed==0.5.1
aiosqlite==0.20.0
httpx==0.27.0
six==1.16.0
PyJWT==2.10.1
bcrypt==4.2.1
uvicorn==0.34.0
watchdog==6.0.0
inflection==0.5.1
aiofiles==24.1.0